requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-allauth==0.47.0
oauthlib==3.1.1
psycopg==3.0.8
PyJWT==2.3.0
python3-openid==3.2.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
